refactor(names): change page's names from "Main" to "MainScreen"; add index.ts for all components and pages
