+++ src/components/PublicRoutes/PublicRoutes.tsx
@@ -1,7 +1,7 @@
 import { Navigate, Outlet } from 'react-router-dom';
 import { AppRoute } from '../../const/pageRoutes';
 
-export default function PublicRoutes() {
+export function PublicRoutes() {
   const hasAccess = false;
 
   return hasAccess ? <Navigate to={AppRoute.Main} /> : <Outlet />;
