+++ src/components/PublicRoutes/index.ts
@@ -1,0 +1,1 @@
+export * from './PublicRoutes';
