--- src/pages/FavoritesPlaces/index.ts
@@ -1,1 +1,0 @@
-export * from './FavoritesPlaces';
