+++ src/pages/NotFoundScreen/index.ts
@@ -1,0 +1,1 @@
+export * from './NotFoundScreen';
