+++ src/pages/OfferScreen/Gallery/Gallery.tsx
