+++ src/pages/OfferScreen/Gallery/index.ts
