+++ src/pages/OfferScreen/NearbyPlaces/NearbyPlaces.tsx
