+++ src/pages/OfferScreen/NearbyPlaces/index.ts
