+++ src/pages/OfferScreen/OfferDetails/AmenitiesList/AmenitiesList.tsx
