+++ src/pages/OfferScreen/OfferDetails/AmenitiesList/index.ts
