+++ src/pages/OfferScreen/OfferDetails/FeaturesList/FeaturesList.tsx
