+++ src/pages/OfferScreen/OfferDetails/FeaturesList/index.ts
