+++ src/pages/OfferScreen/OfferDetails/HostInfo/HostInfo.tsx
