+++ src/pages/OfferScreen/OfferDetails/HostInfo/index.ts
