+++ src/pages/OfferScreen/OfferDetails/OfferDetails.tsx
