+++ src/pages/OfferScreen/OfferDetails/ReviewsSection/ReviewItem/ReviewItem.tsx
