+++ src/pages/OfferScreen/OfferDetails/ReviewsSection/ReviewItem/index.ts
