+++ src/pages/OfferScreen/OfferDetails/ReviewsSection/ReviewsSection.tsx
