+++ src/pages/OfferScreen/OfferDetails/ReviewsSection/index.ts
