+++ src/pages/OfferScreen/OfferDetails/index.ts
