+++ src/pages/OfferScreen/index.ts
@@ -1,0 +1,1 @@
+export * from './OfferScreen';
